backend/requirements.txt
# Web
fastapi
uvicorn[standard]
pydantic
python-dotenv

# DB + migrations
sqlalchemy
psycopg2-binary
alembic

# Security / utils
slowapi
PyJWT
cryptography
requests
cachetools

# LLM: Gemini
google-generativeai
httpx

# Supabase
supabase

# Feedback utils (basic scoring)
numpy
scikit-learn
tqdm

# Tests
pytest
pytest-asyncio
